tests/automation/codebase/test_parser.py
import unittest
from unittest.mock import MagicMock, patch

from sentence_transformers import SentenceTransformer
from tree_sitter import Node, Parser

from seer.automation.codebase.models import DocumentChunk
from seer.automation.codebase.parser import Document, DocumentParser, ParentDeclaration, TempChunk


class TestDocumentParser(unittest.TestCase):
    def setUp(self):
        # Mock the SentenceTransformer and Parser
        self.mock_embedding_model = MagicMock(spec=SentenceTransformer)
        self.mock_parser = MagicMock(spec=Parser)
        self.mock_parser.parse.return_value.root_node = MagicMock(spec=Node)
        self.document_parser = DocumentParser(embedding_model=self.mock_embedding_model)

    def test_document_parser_process_document(self):
        # Test processing of a single document
        mock_document = MagicMock(spec=Document)
        mock_document.text = "import os"
        mock_document.path = "test.py"
        mock_document.repo_id = 1

        expected_chunks = [MagicMock(spec=DocumentChunk)]
        self.document_parser.process_document = MagicMock(return_value=expected_chunks)

        result_chunks = self.document_parser.process_document(mock_document)
        self.document_parser.process_document.assert_called_once_with(mock_document)
        self.assertEqual(result_chunks, expected_chunks)

    def test_document_parser_process_documents(self):
        # Test processing of multiple documents
        mock_documents = [MagicMock(spec=Document) for _ in range(2)]
        expected_chunks = [MagicMock(spec=DocumentChunk), MagicMock(spec=DocumentChunk)]
        self.document_parser.process_documents = MagicMock(return_value=expected_chunks)

        result_chunks = self.document_parser.process_documents(mock_documents)
        self.document_parser.process_documents.assert_called_once_with(mock_documents)
        self.assertEqual(result_chunks, expected_chunks)

    # Additional test cases should be added to cover all methods in DocumentParser class.
    # This includes testing edge cases and ensuring that all branches of the code are tested.
    # Mocking of external dependencies like SentenceTransformer and Parser should be done to isolate tests.
